Interscreen nav - basic UI functionality completed
all of the components of the UI should be working now. this includes:
curseScreens, cursePanels, curseTextboxes, and curseItems.

code needs to be cleaned up and better documented to improve clarity of
function.

there are two test screens; a title screen and a generic application
screen, you can navigate from one screen to the other and back either
via panel items or direct keyboard input.