read changelog.txt and UI_documentation.txt
finally starting to get in the groove of this... things will come
quicker in the next couple of days. account creation is ready to be
interfaced to database code. login is almost there. I spent most of
making the background code/docs that will expedite the development of
subsequent screens/modules